requirements.txt
aniso8601==8.0.0
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.1a1
django-environ==0.4.5
django-graphql-jwt==0.3.1
django-multiselectfield==0.1.12
graphene==3.0b2
graphene-django==3.0.0b1
graphql-core==2.3.2
graphql-relay==3.0.0
gunicorn==20.0.4
promise==2.3
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
Rx==1.6.1
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
whitenoise==5.1.0